inmates: support for different compiler
Some inmates might need a library built using a compiler different from
the one used for building the rest of the hypervisor.
This is, for example, the case of an inmate needing the libc support
provided by a bare-metal toolchain.
This patch introduces two additional environment valiables (i.e.
JAILHOUSE_INMATE_CC and JAILHOUSE_INMATE_LD) which, if set, allow to
build the inmate library and demos using a different compiler.

Signed-off-by: Claudio Scordino <claudio@evidence.eu.com>